fix(health): shared external tag constant, move CityPlanningHub out of Contracts/
Hoists Humans.Base.Constants.HealthCheckTags.External so the six independent
"external" literals (Program.cs + five sections' SectionHealthChecks.cs) can't
drift apart — a typo in any copy would silently exempt that check from the
readiness-probe exclusion it exists to guarantee.

CityPlanningHub is internal now that Shell no longer names it by concrete
type, so it moves out of Contracts/ (the section's public-surface folder)
into Services/ alongside the rest of the section's internal implementation.
