+++ src/Sections/Humans.CityPlanning/Controllers/CityPlanningApiController.cs
@@ -1,7 +1,6 @@
 using Humans.Base.Authorization;
 using System.Text.Json;
 using Humans.Camps.Contracts;
-using Humans.CityPlanning.Contracts;
 using Humans.CityPlanning.Services;
 using Humans.Containers.Contracts;
 using Humans.Base.Extensions;
