+++ src/Sections/Humans.Tickets/SectionHealthChecks.cs
@@ -1,3 +1,4 @@
+using Humans.Base.Constants;
 using Humans.Base.Interfaces;
 using Humans.Tickets.Health;
 using Microsoft.Extensions.Configuration;
@@ -14,6 +15,6 @@ internal sealed class SectionHealthChecks : ISectionHealthChecks
 {
     public void AddHealthChecks(IHealthChecksBuilder builder, IConfiguration configuration)
     {
-        builder.AddCheck<TicketVendorHealthCheck>("ticket-vendor", tags: ["external"]);
+        builder.AddCheck<TicketVendorHealthCheck>("ticket-vendor", tags: [HealthCheckTags.External]);
     }
 }
